Add keyname argument

+++ scripts/generate_aws_keypem.sh
@@ -1,4 +1,12 @@
-keyName=my_workspace_admin
+#!/bin/sh
+
+if [ $# -lt 1 ]; 
+then
+    echo "Require keyname option like ./generate_aws_keypem.sh <keyname>"
+    exit 1
+fi
+
+keyName=$1
 keyPath=$HOME/.ssh/$keyName
 
 rm -f $keyPath $keyPath.pub &> /dev/null
